DONE Task 5

+++ src/controlles/productControlles.js
@@ -1,7 +1,8 @@
 import createHttpError from 'http-errors';
-import { Product } from '../models/products.js';
+import { Product } from '../models/product.js';
 
 export const getAllProducts = async (req, res) => {
+  const { _id: userId } = req.user;
   const {
     page = 1,
     perPage = 10,
@@ -14,6 +15,9 @@ export const getAllProducts = async (req, res) => {
   } = req.query;
   const skip = (page - 1) * perPage;
   const productsQuery = Product.find();
+  if (userId) {
+    productsQuery.where('userId').equals(userId);
+  }
   if (category) {
     productsQuery.where('category').equals(category);
   }
@@ -45,32 +49,40 @@ export const getAllProducts = async (req, res) => {
 };
 
 export const getProductById = async (req, res) => {
+  const { _id: userId } = req.user;
   const { productId } = req.params;
-  const product = await Product.findById(productId);
+  const product = await Product.findById({ _id: productId, userId });
   if (!product) {
     return res.status(404).json({
-      message: `Product by${productId} not find`,
+      message: `Product by id${productId} not found`,
     });
   }
   res.status(200).json(product);
 };
+
 export const createProduct = async (req, res) => {
-  const product = await Product.create(req.body);
+  const { _id: userId } = req.user;
+  const product = await Product.create({ ...req.body, userId });
   res.status(201).json(product);
 };
+
 export const deleteProduct = async (req, res) => {
+  const { _id: userId } = req.user;
   const { productId } = req.params;
-  const product = await Product.findOneAndDelete({ _id: productId });
+  const product = await Product.findOneAndDelete({ _id: productId, userId });
   if (!product) {
     throw createHttpError(404, 'Product not found');
   }
   res.status(200).json(product);
 };
+
 export const updateProducts = async (req, res) => {
+  const { _id: userId } = req.user;
   const { productId } = req.params;
   const product = await Product.findOneAndUpdate(
     {
       _id: productId,
+      userId,
     },
     req.body,
   );
